refactor: extract the npx pin parser out of the MCP server config
